AER-4506: Improve command performance in presence of high latency node
Problem:
--------
When a single node in the cluster is experiencing high latency, running
`asadm -e "info"` may take more than 10 seconds to complete.

Solution:
---------
Replace info requests with Futures so that we can fire
multiple info requests simultaneously and retrieve the results
asynchronously. In addition, for command that call multiple sub
commands, execute sub commands as futures as well. To serialize results
return the rendering of a view as a Future that has not started and
modify the base controller to execute each view serially.